test(tui): harden markdown theme regression probe
